+++ ext/phalcon/mvc/view/engine/volt/exception.zep.h
@@ -1,0 +1,5 @@
+
+extern zend_class_entry *phalcon_mvc_view_engine_volt_exception_ce;
+
+ZEPHIR_INIT_CLASS(Phalcon_Mvc_View_Engine_Volt_Exception);
+